fix stuff I broke and put code where it's suppposed to be
