+++ darwin/urls.py
@@ -3,6 +3,7 @@ from . import views
 
 urlpatterns = [
     path('', views.api_root, name='root'),
+    path('login/', views.CustomAuthToken.as_view()),
     path('ideas/', views.ideas, name='ideas'),
     path('boards/', views.BoardList.as_view(), name='boards'),
     path('boards/<int:pk>/', views.BoardDetail.as_view(), name='board'),
@@ -11,4 +12,4 @@ urlpatterns = [
     path('users/<int:pk>/', views.UserDetail.as_view(), name='user'),
     path('users/<int:user_id>/votes/', views.user_votes, name='user_votes'),
     path('users/<int:idea_id>/ideas/', views.user_ideas, name='user_ideas'),
-]
+]
